requirements.txt
python-dotenv
aioretry
# aiomysql
g4f
aiogram==2.*
aioschedule
motor
pytest
pytest-asyncio
bs4
pillow